+++ docs/tutorials/belref/settings.py
@@ -2,7 +2,7 @@ from lino.projects.belref.settings import *
 
 
 class Site(Site):
-    default_ui = 'extjs'
+    default_ui = 'lino.modlib.extjs'
 
 SITE = Site(globals())